Fixies

+++ app/UI/Web/User/Logout/LogoutPresenter.php
@@ -2,10 +2,12 @@
 
 namespace App\UI\Web\User\Logout;
 
+use App\Core\AutoLoginAuthenticator;
 use App\Model\Entity\UserAutoLoginEntity;
 use App\Model\Entity\UserEntity;
 use Doctrine\DBAL\Exception as DbalException;
 use Nette\Application\UI\Presenter;
+use Nette\Http\IResponse;
 use Nette\Localization\Translator;
 use Nettrine\ORM\EntityManagerDecorator;
 
@@ -20,33 +22,41 @@ class LogoutPresenter extends Presenter
 
     public function actionDefault() : void
     {
-        $userEntity = $this->em
-            ->getRepository(UserEntity::class)
-            ->findOneBy(
-                [
-                    'id' => $this->getUser()->getId(),
-                ]
-            );
-
-        if (!$userEntity) {
-            $this->error('user not found');
+        if (!$this->getUser()->isLoggedIn()) {
+            $this->error('User not logged in', IResponse::S400_BadRequest);
         }
 
-        $userAutoLoginEntity = $this->em
-            ->getRepository(UserAutoLoginEntity::class)
-            ->findOneBy(
-                [
-                    'user' => $userEntity,
-                    'token' => $this->getHttpRequest()->getCookie('autoLogin'),
-                ]
-            );
+        if ($this->getHttpRequest()->getCookie(AutoLoginAuthenticator::COOKIE_NAME) !== null) {
+            $userEntity = $this->em
+                ->getRepository(UserEntity::class)
+                ->findOneBy(
+                    [
+                        'id' => $this->getUser()->getId(),
+                    ]
+                );
 
-        try {
-            $this->em->remove($userAutoLoginEntity);
-            $this->em->flush();
-        } catch (DbalException $exception) {
-            $this->flashMessage($exception->getMessage(), 'danger');
-            $this->redrawControl('flashes');
+            if (!$userEntity) {
+                $this->error('user not found');
+            }
+
+            $userAutoLoginEntity = $this->em
+                ->getRepository(UserAutoLoginEntity::class)
+                ->findOneBy(
+                    [
+                        'user' => $userEntity,
+                        'token' => $this->getHttpRequest()->getCookie(AutoLoginAuthenticator::COOKIE_NAME),
+                    ]
+                );
+
+            if ($userAutoLoginEntity) {
+                try {
+                    $this->em->remove($userAutoLoginEntity);
+                    $this->em->flush();
+                } catch (DbalException $exception) {
+                    $this->flashMessage($exception->getMessage(), 'danger');
+                    $this->redrawControl('flashes');
+                }
+            }
         }
 
         $this->getUser()->logout(true);
